requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apt-xapian-index==0.49
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blinker==1.4
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
galternatives==1.0.8
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.2.0
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lubuntu-update-notifier==0.4.dev0
Markdown==3.3.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
pexpect==4.8.0
Pillow==9.0.1
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
python-apt==2.4.0+ubuntu1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2022.10.31
reportlab==3.6.8
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
tqdm==4.65.0
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
